interpreter: explicitly evaluate value-forms
Due to complex values, there was undesirable non-determinism in the
implementation of the interpreter:

    eval V V :- val V.
    val (variant C Vs) :- constructor C, foreach val Vs.
    eval (variant C Ts) (variant C Vs) :- constructor C, foreach2 eval Ts Vs.

Taken together, these three rules imply that there are two ways to
evaluate a variant that is already a value (for example a variant with
literal arguments): either directly recognize that it is a value using
the (val (variant ...)) rule, or evaluate each subcomponent to a value.

To remove this non-determinism, this patch removes the (eval V V :- val V)
rule, replacing it by base cases for simple values only (not complex values):

    eval (lam R) (lam R).
    eval (lit L) (lit L) :- val_lit L.
    eval N N :- nom N.

However, doing only that is problematic for our (piv x \ Goal)
abstraction form, which introduced a (val x =>) hypothesis in the
search context. With the new rules, we don't have

    pi x \ val x => eval x x

anymore. The solution is to define a new predicate "abstract_value x"
for abstract/indecomposable values, introduced by (piv x), and
a restricted eval-value rule for abstract values:

  piv G :- pi x \ abstract_value x => G x.
  val A :- abstract_value A.
  eval A A :- abstract_value A.