select_clause needs to be turned back eval_clause...
The problem comes from undertermined nominals, as in Dale's example

    match unit with
    | nab X in r @ X -> r @ X

In this example, "X" is not bound to any nominal defined outside the
clause. Previously we returned the right-hand-side of the clause
(here `r @ X`) to the toplevel evaluator to continue evaluation, but
in that context we forgo the information that X is a nominal, and the
evaluation of (r @ X) gets stuck.

We need to perform the evaluation of the right-hand-side within the
scope of the "nab" clause-evaluation rule, which can properly assume
that those introduced names are nominals.

The problem doesn't appear for nab-nominals that are determined to be
equal to an outside nominal, as the returned right-hand-side only
contained references to the outside nominal, known to be a nominal.